requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlmodel>=0.0.22
apscheduler>=3.10,<4
jinja2>=3.1
python-multipart>=0.0.12
pydantic-settings>=2.6
aiosqlite>=0.20
PyYAML>=6.0
